modules: add admin stats module.
Work in progress. Currently only charts a snapshot in time,
so no real time updates... Also more work on the graphs themselves
are needed.

Still missing (line) graphs for request counts for the different
time spans. But that is not really effectful until there's live
update of the graph data, appending the latest statistics.