mod_survey: New question type "multiple choice"
The multiple choice option also has the option that the multiple
choice is a numeric value, in which case an overview of the total
value will be shown in the printable list and beneath the survey pie
chart. This is useful for creating forms which require you to enter an
amount or quantity, e.g. for a reservation system.

One drawback: choice values cannot currently be translated. But in
that case use the thurstone question type.

Also documented the survey question types somewhat.
(cherry picked from commit 181a4935bcb8841104e3ca3d2cc185837311bd2c)