core: catinclude also accepts a list of atoms instead of an id
This makes it in the admin again possible to override the
`_admin_overview_list.tpl` file based on the category that is
selected. This restores the functionality that was removed in commit
087415d1fcebe47cb0ad88eaa1e6510019631975, but now implemented in a
different, more explicit way.

Documentation for the new `m.category[222].is_a` option and the extra
feature of `catinclude` is provided.