core: Fixes to image cropping with a center coordinate
The cropping point is now actually the center point around the crop
coordinate, not the top-left.

When 'crop' does not have any arguments and a resource id is given in
the image tag, it will load the `crop_center` property from the
resource and use this as the cropping center.

Also updated the documentation