mod_base: Websocket handle_message callback now gets the 'From' pid.
Previously it was not possible to send a message back to the websocket
in a custom websocket handler's `websocket_message` function, because
the pid of the websocket sender process was not known. This commit
fixes that by propagating the pid of the send loop into the
handle_message function.

This introduces a backward incompatibility with any custom websocket
handlers.