mod_base: controller_page - check if page path equals the request path
This change lets controller_page check if we need to be at a different
URL when the page path of a resource is set.

If set, we need to redirect there if the current request's path is not
equal to the resource's path. This prevents duplicated content on
multiple URLs.

(cherry picked from commit 18bda757da0cfee42a0621dd8978b0d0d24f3feb)