core/epgsql: more improvements to connection handling.
This continues the changes started in commit e9eeda7cd87f7d00f9554ad76cdaf0f69d8f34d1.

When a new connection is reqeusted from the db pool, it will only put
the client on the waiting queue in case of errors if there already are
other successfull connections.