mod_acl_user_groups: Fix permission issue
When content_group_id is not passed on insert, the insert fails when the
user has no permissions to insert anything in any category *except* a
category where he is a collaboration group of.

With this fix, the content group id will be set to the first
collaboration group where the user has insert rights on.

(cherry picked from commit 4ebe6d17fce7db13941b8e5e60d9eed1343222b5)